Use xarray >=0.15.1 builtin differentiate rather than infinite-diff
WARNING: I changed a decent amount of logic in this commit,
potentially breaking it in places.  Of greatest concern in that
regard is `kuo_el`.  Going to commit this anyways because I
don't need that code for the foreseeable future and so it doesn't
seem worth the time to extensively test.